+++ cognite/neat/_issues/errors/__init__.py
@@ -31,12 +31,7 @@ from ._resources import (
     ResourceNotFoundError,
     ResourceRetrievalError,
 )
-from ._workflow import (
-    WorkflowConfigurationNotSetError,
-    WorkFlowMissingDataError,
-    WorkflowStepNotInitializedError,
-    WorkflowStepOutputError,
-)
+from ._wrapper import MetadataValueError
 
 __all__ = [
     "AuthorizationError",
@@ -47,6 +42,7 @@ __all__ = [
     "FileNotFoundNeatError",
     "FileReadError",
     "FileTypeUnexpectedError",
+    "MetadataValueError",
     "NeatError",
     "NeatImportError",
     "NeatTypeError",
@@ -70,10 +66,6 @@ __all__ = [
     "ResourceRetrievalError",
     "ReversedConnectionNotFeasibleError",
     "RowError",
-    "WorkFlowMissingDataError",
-    "WorkflowConfigurationNotSetError",
-    "WorkflowStepNotInitializedError",
-    "WorkflowStepOutputError",
 ]
 
 _NEAT_ERRORS_BY_NAME = {error.__name__: error for error in _get_subclasses(NeatError, include_base=True)}