+++ cognite/neat/_session/_prepare.py
@@ -3,6 +3,7 @@ from typing import Any
 
 from rdflib import URIRef
 
+from cognite.neat._alpha import AlphaFlags
 from cognite.neat._graph.transformers import (
     ConnectionToLiteral,
     ConvertLiteral,
@@ -12,9 +13,7 @@ from cognite.neat._graph.transformers import (
 from cognite.neat._graph.transformers._rdfpath import MakeConnectionOnExactMatch
 from cognite.neat._issues import IssueList
 from cognite.neat._issues.errors import NeatValueError
-from cognite.neat._rules.transformers import (
-    PrefixEntities,
-)
+from cognite.neat._rules.transformers import PrefixEntities, StandardizeNaming
 from cognite.neat._utils.text import humanize_collection
 
 from ._state import SessionState
@@ -264,3 +263,12 @@ class DataModelPrepareAPI:
         """
 
         return self._state.rule_transform(PrefixEntities(prefix))  # type: ignore[arg-type]
+
+    def standardize_naming(self) -> IssueList:
+        """Standardize the naming of all views/classes/properties in the data model.
+
+        For classes/views/containers, the naming will be standardized to PascalCase.
+        For properties, the naming will be standardized to camelCase.
+        """
+        AlphaFlags.standardize_naming.warn()
+        return self._state.rule_transform(StandardizeNaming())